requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
ca-certificates=2021.1.19=h06a4308_0
certifi=2020.12.5=py38h06a4308_0
ed25519=1.5=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
ncurses=6.2=he6710b0_1
npyscreen=4.10.5=pypi_0
openssl=1.1.1j=h27cfd23_0
pip=20.3.3=py38h06a4308_0
protobuf=3.15.1=pypi_0
python=3.8.5=h7579374_1
pyzmq=20.0.0=py38h2531618_1
readline=8.1=h27cfd23_0
setuptools=52.0.0=py38h06a4308_0
six=1.15.0=pypi_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zlib=1.2.11=h7b6447c_3
